Adding method to grab just URLs for specific range
This will make it easier to queue these items into a larger Rails-type
application.  The larger application could ask for this block of urls,
and then it could handle the looping or instantiate objects in a
resqueue or delayed_job system to do the actual processing.

This gem can focus on the logic of parsing the webpages without worrying
about (precisely) the struture of the database in which that information
will be used.