Updating for new format of COA Order List pages
TAMES now has a new format that no longer splits the case name and the
lower court name with a "--".  That was the signal that had been used to
divide the two by the scraper.  Now, it is done by relying on the span
tags (which OCA is using the bold text).

This commit updates the (old) tests for that change.  It also migrates
the (old) tests for two formerly legacy courts that have now moved over
to the TAMES system from the Legacy system, and it creates a new test
suite targeted at the 13th Court (the last on the Legacy system).

Another small change is adjusting for a URL change on opinions. The
court now has a variable (!) within the scraped URL, which could make it
difficult to revisit the sites later. My code change relies on the user
to pass the COA to the parsing routine, which then uses that information
to do the substitution by hand and code the "real" URL.