Extracting logic about case styles and origins
The new TamesCaseStyle class encapsulates the knowledge of how to
extract the case names and courts of origin (which are in the same table
cell and have varying formatting).

The oldest pages use a punctuation divider ("--") while the newer pages
use bolding to distinguish the (bolded) case name from the (unbolded)
court of origin.  The formatting is in a span, which is used to
distinguish these text blocks.